Total Purchase amount for Sub facilitator issue fixed

+++ additional_addons/kingswood/kw_stock.py
@@ -4662,7 +4662,7 @@ class stock_picking_out(osv.osv):
                     if sub_part_ids:
                         sub_part_ids = sub_part_ids[0]
                         sub_part = sub_part_obj.browse(cr, uid, sub_part_ids)
-                        if sub_part.total_purchase >= float(874575):
+                        if sub_part.total_purchase >= float(1850000):
                             raise osv.except_osv(_('Warning'),_('Total Purcase is exceeded for the selected Sub Facilitator.'))
                     else:
                         raise osv.except_osv(_('Warning'),_('Please create the Sub Facilitator for the Current Year'))
@@ -7529,7 +7529,7 @@ class stock_picking_in(osv.osv):
                     if sub_part_ids:
                         sub_part_ids = sub_part_ids[0]
                         sub_part = sub_part_obj.browse(cr, uid, sub_part_ids)
-                        if sub_part.total_purchase >= float(874575):
+                        if sub_part.total_purchase >= float(1850000):
                             raise osv.except_osv(_('Warning'),_('Total Purcase is exceeded for the selected Sub Facilitator.'))
         return res
 
@@ -7663,7 +7663,7 @@ class stock_picking_in(osv.osv):
                 if sub_part_ids:
                     sub_part_ids = sub_part_ids[0]
                     sub_part = sub_part_obj.browse(cr, uid, sub_part_ids)
-                    if sub_part.total_purchase >= float(874575):
+                    if sub_part.total_purchase >= float(1850000):
                         raise osv.except_osv(_('Warning'),_('Total Purcase is exceeded for the selected Sub Facilitator.'))
         return res
